requirements.txt
asgiref==3.2.3
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
djangorestframework==3.11.0
firebase-admin==3.2.1
gis==0.2.1
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.0
idna==2.9
msgpack==1.0.0
protobuf==3.11.3
psycopg2-binary==2.8.4
pusher-push-notifications==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfcm==1.4.7
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8